More deduping and cleanup
