+++ app/pages/project/access/ProjectAccessPage.tsx
@@ -7,10 +7,9 @@
  */
 import type { LoaderFunctionArgs } from 'react-router'
 
-import { queryClient } from '@oxide/api'
+import { api, q, queryClient } from '@oxide/api'
 import { Access16Icon, Access24Icon } from '@oxide/design-system/icons/react'
 
-import { accessQueries } from '~/api/access-queries'
 import { DocsPopover } from '~/components/DocsPopover'
 import { RouteTabs, Tab } from '~/components/RouteTabs'
 import { getProjectSelector, useProjectSelector } from '~/hooks/use-params'
@@ -19,13 +18,13 @@ import { docLinks } from '~/util/links'
 import { pb } from '~/util/path-builder'
 
 export async function clientLoader({ params }: LoaderFunctionArgs) {
-  const selector = getProjectSelector(params)
+  const { project } = getProjectSelector(params)
   await Promise.all([
-    queryClient.prefetchQuery(accessQueries.siloPolicy()),
-    queryClient.prefetchQuery(accessQueries.projectPolicy(selector)),
+    queryClient.prefetchQuery(q(api.policyView, {})),
+    queryClient.prefetchQuery(q(api.projectPolicyView, { path: { project } })),
     // used to resolve user names
-    queryClient.prefetchQuery(accessQueries.userList()),
-    queryClient.prefetchQuery(accessQueries.groupList()),
+    queryClient.prefetchQuery(q(api.userList, {})),
+    queryClient.prefetchQuery(q(api.groupList, {})),
   ])
   return null
 }
