+++ app/util/access.ts
@@ -11,6 +11,16 @@ import { type BadgeColor } from '@oxide/design-system/ui'
 import type { IdentityType, RoleKey } from '~/api'
 import type { IdentityFilter } from '~/types/access'
 
+/** Filter rows by identity type based on the filter parameter */
+export function filterByIdentityType<T extends { identityType: IdentityType }>(
+  rows: T[],
+  filter: IdentityFilter
+): T[] {
+  if (filter === 'users') return rows.filter((row) => row.identityType === 'silo_user')
+  if (filter === 'groups') return rows.filter((row) => row.identityType === 'silo_group')
+  return rows
+}
+
 export const identityTypeLabel: Record<IdentityType, string> = {
   silo_group: 'Group',
   silo_user: 'User',
