+++ test/e2e/silo-access.e2e.ts
@@ -7,6 +7,7 @@
  */
 import { user3, user4 } from '@oxide/api-mocks'
 
+import { testAddUserOnAllTabAndVerifyOnUsersTabs } from './access-test-helpers'
 import { expect, expectNotVisible, expectRowVisible, expectVisible, test } from './utils'
 
 test('Click through silo access page', async ({ page }) => {
@@ -81,64 +82,9 @@ test('Click through silo access page', async ({ page }) => {
   await expect(user3Row).toBeHidden()
 })
 
-test('Add user on All tab and verify on Users tab', async ({ page }) => {
-  await page.goto('/access/all')
-
-  // Start on the All tab
-  await expectVisible(page, ['role=tab[name="All"][selected]'])
-
-  const allTable = page.locator('table')
-
-  // Add Hans Jonas as viewer
-  await page.click('role=button[name="Add user or group"]')
-  await expectVisible(page, ['role=heading[name*="Add user or group"]'])
-
-  await page.click('role=button[name*="User or group"]')
-  await page.click('role=option[name="Hans Jonas"]')
-  await page.getByRole('radio', { name: /^Viewer / }).click()
-  await page.click('role=button[name="Assign role"]')
-
-  // User shows up in the All tab table
-  await expectRowVisible(allTable, {
-    Name: 'Hans Jonas',
-    Type: 'User',
-    Role: 'silo.viewer',
-  })
-
-  // Navigate to Users tab
-  await page.click('role=tab[name="Users"]')
-  await expectVisible(page, ['role=tab[name="Users"][selected]'])
-
-  // Verify the URL changed to /users
-  await expect(page).toHaveURL(/\/access\/users$/)
-
-  const usersTable = page.locator('table')
-
-  // User should still be visible on Users tab
-  await expectRowVisible(usersTable, {
-    Name: 'Hans Jonas',
-    Role: 'silo.viewer',
-  })
-
-  // Type column should not be present on Users tab (since all are users)
-  await expectNotVisible(page, ['role=columnheader[name="Type"]'])
-
-  // Navigate to Groups tab
-  await page.click('role=tab[name="Groups"]')
-  await expectVisible(page, ['role=tab[name="Groups"][selected]'])
-
-  // Hans Jonas should NOT be visible on Groups tab
-  await expectNotVisible(page, ['role=cell[name="Hans Jonas"]'])
-
-  // Type column should not be present on Groups tab either
-  await expectNotVisible(page, ['role=columnheader[name="Type"]'])
-
-  // Go back to All tab and verify Hans Jonas is still there
-  await page.click('role=tab[name="All"]')
-  await expectVisible(page, ['role=tab[name="All"][selected]'])
-  await expectRowVisible(allTable, {
-    Name: 'Hans Jonas',
-    Type: 'User',
-    Role: 'silo.viewer',
+test('Add silo user on All tab and verify on Users tab', async ({ page }) => {
+  await testAddUserOnAllTabAndVerifyOnUsersTabs(page, {
+    baseUrl: '/access/all',
+    rolePrefix: 'silo',
   })
 })
